requirements.txt
google-cloud-bigquery==3.3.2
pandas==1.3.5
scikit-learn==1.0.2
torch==1.12.1
tensorboard==2.10.1
tqdm==4.64.1
pandas-gbq==0.17.9
grpcio-tools==1.44.0 # for weird protobuf errors
seaborn==0.12.1
jupyter==1.0.0 # likely useful
pyyaml==6.0
requests==2.25.0
xmltodict==0.12.0
azure.cosmos==4.2.0

